requirements.txt
pynubank
ipdb
pandas
cherrypy
Jinja2
python-dotenv